refactor: Ajuste estrutura de pastas do frontend para que seja mais fácil dar manutenção no futuro
