+++ frontend/src/components/Home/LocalCard/LocalCard.vue
