+++ frontend/src/components/Home/LocalInfo/LocalInfo.vue
