+++ frontend/src/components/Layout/ModalConfirmacaoInativacao.vue
