+++ frontend/src/router/index.ts
@@ -1,15 +1,15 @@
 import { createRouter, createWebHistory } from 'vue-router';
-import CadastroUsuario from '../views/CadastroUsuarioView.vue';
-import HomeView        from '../views/HomeView.vue';
-import LocalDetalheView from '../views/LocalDetalheView.vue';
-import AtividadesView from '../views/AtividadesView.vue';
-import InstalacoesView from '../views/InstalacoesView.vue';
-import LogsView from '../views/LogsView.vue';
-import UsuariosView from '../views/UsuariosView.vue';
-import PerfilView from '../views/PerfilView.vue';
-import LoginView from '../views/LoginView.vue';
-import RecuperarSenhaView from '../views/RecuperarSenhaView.vue';
-import NovaSenhaView from '../views/NovaSenhaView.vue';
+import HomeView                           from '@/views/Home/HomeView.vue';
+import CadastroUsuario                    from '@/views/Usuarios/Cadastrar/CadastroUsuarioView.vue';
+import LocalDetalheView                   from '@/views/Local/LocalDetalheView.vue';
+import AtividadesView                     from '@/views/atividades/AtividadesView.vue';
+import InstalacoesView                    from '@/views/Instalacoes/InstalacoesView.vue';
+import LogsView                           from '@/views/Logs/LogsView.vue';
+import UsuariosView                       from '@/views/Usuarios/UsuariosView.vue';
+import PerfilView                         from '@/views/Usuarios/Perfil/PerfilView.vue';
+import LoginView                          from '@/views/Login/LoginView.vue';
+import RecuperarSenhaView                 from '@/views/RecuperarSenha/RecuperarSenhaView.vue';
+import NovaSenhaView                      from '@/views/NovaSenha/NovaSenhaView.vue';
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
@@ -18,19 +18,6 @@ const router = createRouter({
       path: '/',
       name: 'home',
       component: HomeView,
-    },
-    // {
-    //   path: '/about',
-    //   name: 'about',
-    //   // route level code-splitting
-    //   // this generates a separate chunk (About.[hash].js) for this route
-    //   // which is lazy-loaded when the route is visited.
-    //   component: () => import('../views/AboutView.vue'),
-    // },
-    {
-      path: "/cadastrar",
-      name: "cadastroUsuario",
-      component: CadastroUsuario,
     },
     {
       path: '/local/:id',
@@ -59,7 +46,12 @@ const router = createRouter({
       component: UsuariosView,
     },
     {
-      path: '/perfil',
+      path: "/usuarios/cadastrar",
+      name: "cadastroUsuario",
+      component: CadastroUsuario,
+    },
+    {
+      path: '/usuarios/perfil',
       name: 'perfil',
       component: PerfilView,
     },
