--- frontend/src/views/LogsView.vue
@@ -1,9 +1,0 @@
-<template>
-  <PesquisarLocal />
-  <div style="margin-top: 32px;"><LogsTable /></div>
-</template>
-
-<script setup>
-import PesquisarLocal from '@/components/PesquisarLocal.vue';
-import LogsTable from '@/components/LogsTable.vue';
-</script>
