+++ frontend/src/views/RecuperarSenha/RecuperarSenhaView.vue
@@ -33,7 +33,7 @@
 
 <script setup lang="ts">
   
-  import Loader                       from '@/components/Loader.vue';
+  import Loader                       from '@/components/Layout/Loader.vue';
   import InputText                    from 'primevue/inputtext';
   import FloatLabel                   from 'primevue/floatlabel';
   import InputMask                    from 'primevue/inputmask';
